@@ -12,41 +12,43 @@
     "lint:fix": "eslint './src/**/*.{js,ts,tsx,jsx}' --fix"
   },
   "dependencies": {
-    "@reduxjs/toolkit": "^1.6.0",
-    "ahooks": "^2.10.6",
-    "antd": "^4.16.3",
-    "axios": "^0.21.1",
+    "@reduxjs/toolkit": "^1.9.2",
+    "ahooks": "^3.7.4",
+    "antd": "^5.2.1",
+    "axios": "^1.3.3",
+    "history": "^5.3.0",
     "md5": "^2.3.0",
     "nprogress": "^0.2.0",
-    "qs": "^6.10.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.4",
-    "react-router-dom": "^5.2.0",
+    "qs": "^6.11.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.8.1",
     "react-router-redux": "^4.0.8",
-    "react-transition-group": "^4.4.2",
-    "redux": "^4.1.0",
+    "react-transition-group": "^4.4.5",
+    "redux": "^4.2.1",
     "redux-react-hook": "^4.0.3",
-    "store2": "^2.12.0",
+    "store2": "^2.14.2",
     "toastr": "^2.1.4"
   },
   "devDependencies": {
-    "@babel/core": "^7.14.6",
-    "@babel/plugin-proposal-decorators": "^7.14.5",
-    "@babel/preset-env": "^7.14.5",
-    "@rollup/plugin-babel": "^5.3.0",
-    "@vitejs/plugin-react-refresh": "^1.3.3",
-    "autoprefixer": "^10.2.6",
-    "eslint-config-lcy-react": "^3.1.2",
-    "husky": "^6.0.0",
-    "less": "^4.1.1",
-    "lint-staged": "^11.0.0",
-    "postcss": "^8.3.4",
-    "sass": "^1.35.1",
-    "vite": "^2.3.7",
-    "vite-plugin-pwa": "^0.8.1",
-    "vite-plugin-style-import": "^1.0.0",
-    "vite-plugin-windicss": "^1.0.4"
+    "@babel/core": "^7.20.12",
+    "@babel/plugin-proposal-decorators": "^7.20.13",
+    "@babel/preset-env": "^7.20.2",
+    "@rollup/plugin-babel": "^6.0.3",
+    "@vitejs/plugin-react-refresh": "^1.3.6",
+    "autoprefixer": "^10.4.13",
+    "consola": "^2.15.3",
+    "eslint-config-lcy-react": "^3.4.0",
+    "husky": "^8.0.3",
+    "less": "^4.1.3",
+    "lint-staged": "^13.1.2",
+    "postcss": "^8.4.21",
+    "sass": "^1.58.1",
+    "vite": "^4.1.1",
+    "vite-plugin-pwa": "^0.14.4",
+    "vite-plugin-style-import": "^2.0.0",
+    "vite-plugin-windicss": "^1.8.10"
   },
   "lint-staged": {
     "*.{js,ts,tsx,jsx}": [
