@@ -1,12 +1,1 @@
-module.exports = {
-    presets: [
-        [
-            '@babel/preset-env',
-            {
-                useBuiltIns: 'usage',
-                corejs: '3'
-            }
-        ]
-    ],
-    plugins: ['@babel/plugin-proposal-class-properties']
-}
+module.exports = {}
