@@ -1,10 +1,11 @@
 import React, { useRef } from 'react'
-import { withRouter } from 'react-router-dom'
+import { useLocation } from 'react-router-dom'
 import SignIn from '@/components/sign-in.jsx'
 import SignUp from '@/components/sign-up.jsx'
 
-const Sign = props => {
-    const backend = useRef(props.location.pathname.indexOf('/backend') === 0)
+const Sign = () => {
+    const location = useLocation()
+    const backend = useRef(location.pathname.indexOf('/backend') === 0)
     const signUpHtml = !backend.current ? <SignUp /> : ''
     const signInHtml = !backend.current ? <SignIn /> : ''
     return (
@@ -14,4 +15,4 @@ const Sign = props => {
         </>
     )
 }
-export default withRouter(Sign)
+export default Sign
