@@ -1,21 +1,20 @@
 import React from 'react'
-import { render } from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import { Provider } from 'react-redux'
 import { StoreContext } from 'redux-react-hook'
-import { BrowserRouter as Router } from 'react-router-dom'
+import { BrowserRouter } from 'react-router-dom'
 
 import Root from './pages/app'
 import store from './store'
 
 console.log('当前环境: ' + import.meta.env.VITE_APP_ENV)
 
-render(
+createRoot(document.getElementById('root')).render(
     <Provider store={store}>
         <StoreContext.Provider value={store}>
-            <Router>
+            <BrowserRouter>
                 <Root />
-            </Router>
+            </BrowserRouter>
         </StoreContext.Provider>
-    </Provider>,
-    document.getElementById('root')
+    </Provider>
 )
